pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "max-cli"
version = "0.1.0"
description = "The AI-powered CLI companion for high-performance tasks."
authors = [{name = "Abubakr Alsheikh", email = "abubakralsheikh@outlook.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "pymupdf>=1.22.0",
    "pydantic-settings>=2.0.0",
    "openai>=1.0.0" 
]

[project.scripts]
max = "max_cli.main:app"

[tool.ruff]
line-length = 88
